requirements.txt
pandas
numpy
seaborn
sklearn
statsmodels
flask
flask_cors
gunicorn
mysql-connector-python==8.0.22